All necessary functions according to the technical specifications have been added

+++ src/main/java/org/example/AccountProperties.java
@@ -2,24 +2,25 @@ package org.example;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import java.math.BigDecimal;
 
 @Component
 public class AccountProperties {
 
-    private final float default_amount;
-    private final float transfer_commission;
+    private final BigDecimal default_amount;
+    private final BigDecimal transfer_commission;
 
-    public AccountProperties(@Value("${account.default-amount}") float default_amount,
-                             @Value("${account.transfer-commission}") float transfer_commission
-    ){this.default_amount=default_amount;
-      this.transfer_commission=transfer_commission;
+    public AccountProperties(@Value("${account.default-amount}") BigDecimal default_amount,
+                             @Value("${account.transfer-commission}") BigDecimal transfer_commission) {
+        this.default_amount = default_amount;
+        this.transfer_commission = transfer_commission;
     }
 
-    public float getDefault_amount() {
+    public BigDecimal getDefault_amount() {
         return default_amount;
     }
 
-    public float getTransfer_commission() {
+    public BigDecimal getTransfer_commission() {
         return transfer_commission;
     }
 }
